all resolvers tested

+++ server/schemas/resolvers.js
@@ -58,6 +58,54 @@ const resolvers = {
       }
     },
     // agith
+    getFriendRequest: async (parent, { requestId }, context) => {
+      try {
+        // if (context.user) {
+
+        // }
+        // throw AuthenticationError;
+
+        // dev code
+
+        const friendRequest = await FriendRequest.findById(requestId);
+        // const friendRequest = await FriendRequest.findOne({"_id": requestId}, '_id');
+
+        // console.log(friendRequest, '---------');
+
+        if (!friendRequest) {
+          throw RequestNotFoundError;
+        };
+
+        return friendRequest;
+      } catch (error) {
+        console.log(error);
+        throw error;
+      }
+    },
+    // agith
+    getFriendRequests: async (parent, { username }, context) => {
+      try {
+        // if (context.user) {
+
+        // }
+        // throw AuthenticationError;
+
+        // dev code
+        const usersRequests = await User.findOne(
+          { username },
+          '_id username friendRequests',).populate('friendRequests');
+
+        if (!usersRequests) {
+          throw UserNotFoundError;
+        };
+
+        return usersRequests.friendRequests;
+      } catch (error) {
+        console.log(error);
+        throw error;
+      }
+    },
+    // agith
     getPost: async (parent, { postId }, context) => {
       try {
         // if (context.user) {
@@ -66,7 +114,7 @@ const resolvers = {
         // throw AuthenticationError;
 
         // dev code
-        const post = Post.findById(postId);
+        const post = await Post.findById(postId);
 
         if (!post) {
           throw PostNotFoundError;
@@ -117,7 +165,7 @@ const resolvers = {
           throw PostNotFoundError;
         }
 
-        return post.comments.find((comment) => comment._id==commentId) || {};
+        return post.comments.find((comment) => comment._id == commentId) || {};
       } catch (error) {
         console.log(error);
         throw error;
@@ -265,6 +313,26 @@ const resolvers = {
         throw error;
       }
     },
+    // agith
+    getAllRequests: async (parent, { }, context) => {
+      try {
+        // if (context.user) {
+
+        // }
+        // throw AuthenticationError;
+
+        // dev code
+
+        const friendRequest = await FriendRequest.find();
+        if (!friendRequest) {
+          throw RequestNotFoundError;
+        };
+        return friendRequest;
+      } catch (error) {
+        console.log(error);
+        throw error;
+      }
+    },
   },
   Mutation: {
     // agith
@@ -351,7 +419,8 @@ const resolvers = {
         throw error;
       }
     },
-    deleteUser: async (parent, { userId, }, context) => {
+    // agith
+    deleteUser: async (parent, { userId, password}, context) => {
       try {
         // if (context.user) {
 
@@ -360,35 +429,29 @@ const resolvers = {
 
         // dev code
         const user = await User.findOneAndDelete({ _id: userId });
-
         if (!user) {
           throw UserNotFoundError;
         };
 
-        const posts = await Post.deleteMany({ _id: { $in: user.posts } });
+        const correctPw = await user.isCorrectPassword(password);
+        if (!correctPw) {
+          throw UserNotFoundError;
+        };
 
-        const commentsToDelete = posts.map((post) => {
-          return post.comments
-        }).reduce((accumulator, currentArray) => accumulator.concat(currentArray), []);
+        await Post.deleteMany({ _id: { $in: user.posts } });
 
-        await Comment.deleteMany({
-          $or: [
-            { _id: { $in: commentsToDelete } },
-            { _id: { $in: user.comments } },
-          ]
-        });
+        // await Notification.deleteMany({ _id: { $in: user.notifications } });
 
-        await Notification.deleteMany({ _id: { $in: user.notifications } });
         // need to remove this friend from other friends' lists
-        await User.updateMany(
+        await User.updateMany( // this could be used for removing comments maybe
           { _id: { $in: user.friends } },
           { $pull: { friends: user._id } },
         );
 
-        await Chat.updateMany(
-          { _id: { $in: user.chats } },
-          { $pull: { friends: user._id } },
-        );
+        // await Chat.updateMany(
+        //   { _id: { $in: user.chats } },
+        //   { $pull: { friends: user._id } },
+        // );
 
         return user;
       } catch (error) {
@@ -396,6 +459,7 @@ const resolvers = {
         throw error;
       }
     },
+    // agith
     requestFriend: async (parent, { requesterId, targetId }, context) => {
       try {
         // if (context.user) {
@@ -409,6 +473,8 @@ const resolvers = {
           { $set: { requesterId, targetId } },
           { upsert: true, new: true, setDefaultsOnInsert: true, runValidators: true },
         );
+
+        // const friendRequest = await FriendRequest.create({ requesterId, targetId });
 
         const user = await User.findOneAndUpdate(
           { _id: targetId },
@@ -430,6 +496,7 @@ const resolvers = {
         throw error;
       }
     },
+    // agith
     denyFriend: async (parent, { requestId }, context) => {
       try {
         // if (context.user) {
@@ -438,14 +505,14 @@ const resolvers = {
         // throw AuthenticationError;
 
         // dev code
-        const friendRequest = await FriendRequest.deleteOne({ _id: requestId });
+        const friendRequest = await FriendRequest.findOneAndDelete({ _id: requestId });
 
         if (!friendRequest) {
           return RequestNotFoundError;
         }
 
         const user = await User.findOneAndUpdate(
-          { _id: targetId },
+          { _id: friendRequest.targetId },
           { $pull: { friendRequests: friendRequest._id } },
           {
             new: true,
@@ -463,6 +530,7 @@ const resolvers = {
         throw error;
       }
     },
+    // agith
     acceptFriend: async (parent, { requestId }, context) => {
       try {
         // if (context.user) {
@@ -471,16 +539,21 @@ const resolvers = {
         // throw AuthenticationError;
 
         // dev code
-        const friendRequest = await FriendRequest.deleteOne({ _id: requestId });
+        const friendRequest = await FriendRequest.findOneAndDelete({ _id: requestId });
+        // console.log(friendRequest);
 
         const newFriend = await User.findOneAndUpdate(
           { _id: friendRequest.targetId },
-          { $addToSet: { friends: friendRequest.requesterId } },
+          {
+            $addToSet: { friends: friendRequest.requesterId },
+            $pull: { friendRequests: requestId }
+          },
           {
             new: true,
             runValidators: true
           }
         )
+        // console.log('first');
         if (!newFriend) {
           throw UserNotFoundError;
         };
@@ -493,6 +566,7 @@ const resolvers = {
             runValidators: true
           }
         );
+        // console.log('second');
         if (!user) {
           throw UserNotFoundError;
         };
@@ -502,6 +576,7 @@ const resolvers = {
         throw error;
       }
     },
+    // agith
     removeFriend: async (parent, { me, friend }, context) => {
       try {
         // if (context.user) {
@@ -534,6 +609,42 @@ const resolvers = {
           throw UserNotFoundError;
         };
         return oldFriend;
+      } catch (error) {
+        console.log(error);
+        throw error;
+      }
+    },
+    // agith
+    deleteRequest: async (parent, { requestId }, context) => {
+      try {
+        // if (context.user) {
+
+        // }
+        // throw AuthenticationError;
+
+        // dev code
+        const friendRequest = await FriendRequest.findByIdAndDelete(requestId);
+
+        if (!friendRequest) {
+          throw RequestNotFoundError;
+        };
+
+        const user = await User.findOneAndUpdate(
+          { _id: friendRequest.targetId },
+          {
+            $pull: { friendRequests: requestId }
+          },
+          {
+            new: true,
+            runValidators: true
+          }
+        )
+        if (!user) {
+          throw UserNotFoundError;
+        };
+
+        return friendRequest;
+
       } catch (error) {
         console.log(error);
         throw error;
@@ -647,7 +758,7 @@ const resolvers = {
         // dev code
         const post = await Post.findOneAndUpdate(
           { _id: postId },
-          { $pull: { comments: { _id: commentId} } },
+          { $pull: { comments: { _id: commentId } } },
           {
             // new: true,
             // runValidators: true
